refactor: move ExtractNestedBytes tests from JsonTreeNodeHelperTests to JsonByteExtractorTests
